m2/workshop/ex2/pyproject.toml
[tool.poetry]
name = "bank_api"
version = "0.1.0"
description = ""
authors = ["Robert Owen <robert.owen@softwire.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask-restx = "1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
